project_root/backend/requirements.txt
flask
numpy
pandas
sqlalchemy 
scikit-learn
psycopg2-binary 
python-dotenv 
alembic 
tenacity
celery
redis
pydantic
PyJWT>=2.6,<3.0 
flask_jwt_extended>=4.5
flask-limiter

